Refactored SteamWorksController.
